refac: introduce op-succinct-proof-utils crate
