+++ fault-proof/Cargo.toml
@@ -23,7 +23,7 @@ path = "bin/challenger.rs"
 # local
 op-succinct-client-utils.workspace = true
 op-succinct-host-utils.workspace = true
-op-succinct-ethereum-utils.workspace = true
+op-succinct-proof-utils.workspace = true
 
 # sp1
 sp1-sdk.workspace = true
