+++ fault-proof/src/proposer.rs
@@ -20,9 +20,10 @@ use crate::{
 };
 use op_succinct_client_utils::boot::BootInfoStruct;
 use op_succinct_host_utils::{
-    fetcher::OPSuccinctDataFetcher, get_agg_proof_stdin, get_proof_stdin, get_range_elf_embedded,
-    host::OPSuccinctHost, metrics::MetricsGauge, AGGREGATION_ELF,
+    fetcher::OPSuccinctDataFetcher, get_agg_proof_stdin, get_proof_stdin, host::OPSuccinctHost,
+    metrics::MetricsGauge,
 };
+use op_succinct_proof_utils::{get_range_elf_embedded, AGGREGATION_ELF};
 
 struct SP1Prover {
     network_prover: Arc<NetworkProver>,
