+++ scripts/prove/Cargo.toml
@@ -29,8 +29,8 @@ tracing.workspace = true
 # local
 op-succinct-host-utils.workspace = true
 op-succinct-client-utils.workspace = true
+op-succinct-proof-utils.workspace = true
 op-succinct-scripts = { path = "../utils" }
-op-succinct-ethereum-utils.workspace = true
 
 # alloy
 alloy-primitives.workspace = true
