+++ scripts/utils/bin/gen_sp1_test_artifacts.rs
@@ -2,13 +2,13 @@ use anyhow::Result;
 use clap::Parser;
 use futures::StreamExt;
 use log::info;
-use op_succinct_ethereum_utils::host::initialize_host;
 use op_succinct_host_utils::{
     block_range::{get_validated_block_range, split_range_basic},
     fetcher::OPSuccinctDataFetcher,
-    get_proof_stdin, get_range_elf_embedded,
+    get_proof_stdin,
     host::OPSuccinctHost,
 };
+use op_succinct_proof_utils::{get_range_elf_embedded, initialize_host};
 use op_succinct_scripts::HostExecutorArgs;
 use sp1_sdk::utils;
 use std::{
