+++ utils/client/Cargo.toml
@@ -29,9 +29,6 @@ kona-rpc.workspace = true
 kona-protocol.workspace = true
 kona-genesis.workspace = true
 
-# hana
-hana-oracle = { workspace = true, optional = true }
-
 # general
 anyhow.workspace = true
 async-trait.workspace = true
@@ -44,7 +41,3 @@ tracing.workspace = true
 # kzg
 kzg-rs.workspace = true
 cfg-if.workspace = true
-
-[features]
-default = []
-celestia = ["hana-oracle"]
