+++ utils/host/Cargo.toml
@@ -56,7 +56,3 @@ serde_json.workspace = true
 strum.workspace = true
 tokio.workspace = true
 tracing.workspace = true
-
-[features]
-default = []
-celestia = []
