+++ validity/Cargo.toml
@@ -21,7 +21,7 @@ tracing.workspace = true
 
 # local
 op-succinct-host-utils.workspace = true
-op-succinct-ethereum-utils.workspace = true
+op-succinct-proof-utils.workspace = true
 
 # sp1
 sp1-sdk.workspace = true
@@ -64,7 +64,3 @@ strum_macros.workspace = true
 
 [build-dependencies]
 op-succinct-build-utils.workspace = true
-
-[features]
-# Building without --features will use `SingleChainOPSuccinctHost` with Ethereum DA.
-default = []
